% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposort-package.R
\docType{package}
\name{toposort-package}
\alias{toposort}
\alias{toposort-package}
\title{toposort: Topological Sorting Algorithms}
\description{
Flexible and ergonomic topological sorting implementation for R. Supports a variety of input data encoding (lists of edges or adjacency matrices, graphs edge direction), stable sort variants as well as cycle detection with detailed diagnosis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tzakharko/toposort}
  \item Report bugs at \url{https://github.com/tzakharko/toposort/issues}
}

}
\author{
\strong{Maintainer}: Taras Zakharko \email{taras.zakharko@gmail.com} (\href{https://orcid.org/0000-0001-7601-8424}{ORCID})

}
\keyword{internal}
